edit: moved partnerships to client page
